5004: Refactor tidies
