+++ java/clients/src/main/java/sleeper/clients/admin/AdminClientTrackerFactory.java
@@ -17,8 +17,8 @@ package sleeper.clients.admin;
 
 import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
 
-import sleeper.compaction.trackerv2.job.CompactionJobTrackerFactory;
-import sleeper.compaction.trackerv2.task.CompactionTaskTrackerFactory;
+import sleeper.compaction.tracker.job.CompactionJobTrackerFactory;
+import sleeper.compaction.tracker.task.CompactionTaskTrackerFactory;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.properties.table.TablePropertiesProvider;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
