+++ java/clients/src/main/java/sleeper/clients/report/CompactionTaskStatusReport.java
@@ -21,7 +21,7 @@ import software.amazon.awssdk.services.s3.S3Client;
 import sleeper.clients.report.compaction.task.CompactionTaskQuery;
 import sleeper.clients.report.compaction.task.CompactionTaskStatusReportArguments;
 import sleeper.clients.report.compaction.task.CompactionTaskStatusReporter;
-import sleeper.compaction.trackerv2.task.CompactionTaskTrackerFactory;
+import sleeper.compaction.tracker.task.CompactionTaskTrackerFactory;
 import sleeper.configurationv2.properties.S3InstanceProperties;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.tracker.compaction.task.CompactionTaskTracker;
