+++ java/compaction/compaction-tracker/src/main/java/sleeper/compaction/tracker/job/DynamoDBCompactionJobTracker.java
@@ -27,7 +27,7 @@ import software.amazon.awssdk.services.dynamodb.model.ReturnConsumedCapacity;
 import software.amazon.awssdk.services.dynamodb.model.UpdateItemRequest;
 import software.amazon.awssdk.services.dynamodb.model.UpdateItemResponse;
 
-import sleeper.compaction.trackerv2.CompactionTrackerException;
+import sleeper.compaction.tracker.CompactionTrackerException;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
 import sleeper.core.tracker.compaction.job.query.CompactionJobStatus;
@@ -46,13 +46,13 @@ import java.util.Optional;
 import java.util.function.Supplier;
 import java.util.stream.Stream;
 
-import static sleeper.compaction.trackerv2.job.DynamoDBCompactionJobStatusFormat.UPDATE_TIME;
-import static sleeper.compaction.trackerv2.job.DynamoDBCompactionJobStatusFormat.createJobCommittedUpdate;
-import static sleeper.compaction.trackerv2.job.DynamoDBCompactionJobStatusFormat.createJobCreated;
-import static sleeper.compaction.trackerv2.job.DynamoDBCompactionJobStatusFormat.createJobFailedUpdate;
-import static sleeper.compaction.trackerv2.job.DynamoDBCompactionJobStatusFormat.createJobFinishedUpdate;
-import static sleeper.compaction.trackerv2.job.DynamoDBCompactionJobStatusFormat.createJobStartedUpdate;
-import static sleeper.compaction.trackerv2.task.DynamoDBCompactionTaskStatusFormat.UPDATE_TYPE;
+import static sleeper.compaction.tracker.job.DynamoDBCompactionJobStatusFormat.UPDATE_TIME;
+import static sleeper.compaction.tracker.job.DynamoDBCompactionJobStatusFormat.createJobCommittedUpdate;
+import static sleeper.compaction.tracker.job.DynamoDBCompactionJobStatusFormat.createJobCreated;
+import static sleeper.compaction.tracker.job.DynamoDBCompactionJobStatusFormat.createJobFailedUpdate;
+import static sleeper.compaction.tracker.job.DynamoDBCompactionJobStatusFormat.createJobFinishedUpdate;
+import static sleeper.compaction.tracker.job.DynamoDBCompactionJobStatusFormat.createJobStartedUpdate;
+import static sleeper.compaction.tracker.task.DynamoDBCompactionTaskStatusFormat.UPDATE_TYPE;
 import static sleeper.core.properties.instance.CommonProperty.ID;
 import static sleeper.core.properties.instance.CompactionProperty.COMPACTION_JOB_STATUS_TTL_IN_SECONDS;
 import static sleeper.dynamodb.toolsv2.DynamoDBAttributes.createStringAttribute;
