+++ java/compaction/compaction-tracker/src/test/java/sleeper/compaction/tracker/job/QueryCompactionJobStatusByIdIT.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.trackerv2.job;
+package sleeper.compaction.tracker.job;
 
 import org.junit.jupiter.api.Test;
 
 import sleeper.compaction.core.job.CompactionJob;
-import sleeper.compaction.trackerv2.testutils.DynamoDBCompactionJobTrackerTestBase;
+import sleeper.compaction.tracker.testutils.DynamoDBCompactionJobTrackerTestBase;
 import sleeper.core.partition.Partition;
 import sleeper.core.statestore.FileReferenceFactory;
 
@@ -26,11 +26,12 @@ import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.compaction.core.job.CompactionJobStatusFromJobTestData.compactionJobCreated;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.finishedCompactionRun;
 
-public class QueryAllCompactionJobsIT extends DynamoDBCompactionJobTrackerTestBase {
+public class QueryCompactionJobStatusByIdIT extends DynamoDBCompactionJobTrackerTestBase {
 
     @Test
-    public void shouldReturnMultipleCompactionJobsSortedMostRecentFirst() {
+    public void shouldReturnCompactionJobById() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
@@ -40,46 +41,41 @@ public class QueryAllCompactionJobsIT extends DynamoDBCompactionJobTrackerTestBa
         CompactionJob job2 = jobFactory.createCompactionJob(
                 List.of(fileFactory.rootFile("file2", 456L)),
                 partition.getId());
-        CompactionJob job3 = jobFactory.createCompactionJob(
-                List.of(fileFactory.rootFile("file3", 789L)),
-                partition.getId());
-
-        // When
-        storeJobsCreated(job1, job2, job3);
-
-        // Then
-        assertThat(tracker.getAllJobs(tableId))
-                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(
-                        compactionJobCreated(job3, ignoredUpdateTime()),
-                        compactionJobCreated(job2, ignoredUpdateTime()),
-                        compactionJobCreated(job1, ignoredUpdateTime()));
-    }
-
-    @Test
-    public void shouldExcludeCompactionJobInOtherTable() {
-        // Given
-        Partition partition = singlePartition();
-        FileReferenceFactory fileFactory = fileFactory(partition);
-        CompactionJob job1 = jobFactory.createCompactionJob(
-                List.of(fileFactory.rootFile("file1", 123L)),
-                partition.getId());
-        CompactionJob job2 = jobFactoryForOtherTable().createCompactionJob(
-                List.of(fileFactory.rootFile("file2", 456L)),
-                partition.getId());
 
         // When
         storeJobsCreated(job1, job2);
 
         // Then
-        assertThat(tracker.getAllJobs(tableId))
-                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(compactionJobCreated(job1, ignoredUpdateTime()));
+        assertThat(getJobStatus(job1.getId()))
+                .usingRecursiveComparison(IGNORE_UPDATE_TIMES)
+                .isEqualTo(compactionJobCreated(job1, ignoredUpdateTime()));
     }
 
     @Test
-    public void shouldReturnNoCompactionJobs() {
+    public void shouldReturnFinishedCompactionJobById() {
+        // Given
+        Partition partition = singlePartition();
+        FileReferenceFactory fileFactory = fileFactory(partition);
+        CompactionJob job = jobFactory.createCompactionJob(
+                List.of(fileFactory.rootFile("file", 123L)),
+                partition.getId());
+
+        // When
+        storeJobCreated(job);
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId("test-task").jobRunId("test-run").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId("test-task").jobRunId("test-run").build());
+        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime()).taskId("test-task").jobRunId("test-run").build());
+
+        // Then
+        assertThat(getJobStatus(job.getId()))
+                .usingRecursiveComparison(IGNORE_UPDATE_TIMES)
+                .isEqualTo(compactionJobCreated(job, ignoredUpdateTime(),
+                        finishedCompactionRun("test-task", defaultSummary(), defaultCommitTime())));
+    }
+
+    @Test
+    public void shouldReturnNoCompactionJobById() {
         // When / Then
-        assertThat(tracker.getAllJobs(tableId)).isEmpty();
+        assertThat(tracker.getJob("not-present")).isNotPresent();
     }
 }
