+++ java/compaction/compaction-tracker/src/test/java/sleeper/compaction/tracker/job/QueryCompactionJobStatusUnfinishedIT.java
@@ -13,26 +13,27 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.trackerv2.job;
+package sleeper.compaction.tracker.job;
 
 import org.junit.jupiter.api.Test;
 
 import sleeper.compaction.core.job.CompactionJob;
-import sleeper.compaction.trackerv2.testutils.DynamoDBCompactionJobTrackerTestBase;
+import sleeper.compaction.tracker.testutils.DynamoDBCompactionJobTrackerTestBase;
 import sleeper.core.partition.Partition;
 import sleeper.core.statestore.FileReferenceFactory;
 
-import java.time.Instant;
-import java.time.Period;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.compaction.core.job.CompactionJobStatusFromJobTestData.compactionJobCreated;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.finishedCompactionRun;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.startedCompactionRun;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.uncommittedCompactionRun;
 
-public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobTrackerTestBase {
+public class QueryCompactionJobStatusUnfinishedIT extends DynamoDBCompactionJobTrackerTestBase {
 
     @Test
-    public void shouldReturnCompactionJobsInPeriod() {
+    public void shouldReturnUnfinishedCompactionJobs() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
@@ -47,9 +48,7 @@ public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobTra
         storeJobsCreated(job1, job2);
 
         // Then
-        Instant epochStart = Instant.ofEpochMilli(0);
-        Instant farFuture = epochStart.plus(Period.ofDays(999999999));
-        assertThat(tracker.getJobsInTimePeriod(tableId, epochStart, farFuture))
+        assertThat(tracker.getUnfinishedJobs(tableId))
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
                 .containsExactly(
                         compactionJobCreated(job2, ignoredUpdateTime()),
@@ -57,21 +56,49 @@ public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobTra
     }
 
     @Test
-    public void shouldExcludeCompactionJobOutsidePeriod() {
+    public void shouldIncludeUncommittedCompactionJob() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
         CompactionJob job = jobFactory.createCompactionJob(
-                List.of(fileFactory.rootFile(123L)),
+                List.of(fileFactory.rootFile("file1", 456L)),
                 partition.getId());
 
         // When
         storeJobCreated(job);
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
 
         // Then
-        Instant periodStart = Instant.now().minus(Period.ofDays(2));
-        Instant periodEnd = periodStart.plus(Period.ofDays(1));
-        assertThat(tracker.getJobsInTimePeriod(tableId, periodStart, periodEnd)).isEmpty();
+        assertThat(tracker.getUnfinishedJobs(tableId))
+                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
+                .containsExactly(
+                        compactionJobCreated(job, ignoredUpdateTime(),
+                                uncommittedCompactionRun(DEFAULT_TASK_ID, defaultSummary())));
+    }
+
+    @Test
+    public void shouldExcludeFinishedCompactionJob() {
+        // Given
+        Partition partition = singlePartition();
+        FileReferenceFactory fileFactory = fileFactory(partition);
+        CompactionJob job1 = jobFactory.createCompactionJob(
+                List.of(fileFactory.rootFile("file1", 123L)),
+                partition.getId());
+        CompactionJob job2 = jobFactory.createCompactionJob(
+                List.of(fileFactory.rootFile("file2", 456L)),
+                partition.getId());
+
+        // When
+        storeJobsCreated(job1, job2);
+        tracker.jobStarted(job2.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobFinished(job2.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobCommitted(job2.committedEventBuilder(defaultCommitTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+
+        // Then
+        assertThat(tracker.getUnfinishedJobs(tableId))
+                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
+                .containsExactly(compactionJobCreated(job1, ignoredUpdateTime()));
     }
 
     @Test
@@ -90,60 +117,32 @@ public class QueryCompactionJobStatusByPeriodIT extends DynamoDBCompactionJobTra
         storeJobsCreated(job1, job2);
 
         // Then
-        Instant epochStart = Instant.ofEpochMilli(0);
-        Instant farFuture = epochStart.plus(Period.ofDays(999999999));
-        assertThat(tracker.getJobsInTimePeriod(tableId, epochStart, farFuture))
+        assertThat(tracker.getUnfinishedJobs(tableId))
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
                 .containsExactly(compactionJobCreated(job1, ignoredUpdateTime()));
     }
 
     @Test
-    public void shouldIncludeFinishedStatusUpdateOutsidePeriod() throws Exception {
+    public void shouldIncludeUnfinishedCompactionJobWithOneFinishedRun() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
         CompactionJob job = jobFactory.createCompactionJob(
-                List.of(fileFactory.rootFile(123L)),
-                partition.getId());
-
-        // When
-        Instant periodStart = Instant.now().minus(Period.ofDays(1));
-        storeJobCreated(job);
-        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        Thread.sleep(1);
-        Instant periodEnd = Instant.now();
-        Thread.sleep(1);
-        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-
-        // Then
-        assertThat(tracker.getJobsInTimePeriod(tableId, periodStart, periodEnd))
-                .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(finishedThenCommittedStatusWithDefaults(job));
-    }
-
-    @Test
-    public void shouldIncludeJobCreatedOutsidePeriod() throws Exception {
-        // Given
-        Partition partition = singlePartition();
-        FileReferenceFactory fileFactory = fileFactory(partition);
-        CompactionJob job = jobFactory.createCompactionJob(
-                List.of(fileFactory.rootFile(123L)),
+                List.of(fileFactory.rootFile("file1", 123L)),
                 partition.getId());
 
         // When
         storeJobCreated(job);
-        Thread.sleep(1);
-        Instant periodStart = Instant.now();
-        Thread.sleep(1);
-        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        Instant periodEnd = periodStart.plus(Period.ofDays(1));
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("run-1").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("run-1").build());
+        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime()).taskId(DEFAULT_TASK_ID).jobRunId("run-1").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("run-2").build());
 
         // Then
-        assertThat(tracker.getJobsInTimePeriod(tableId, periodStart, periodEnd))
+        assertThat(tracker.getUnfinishedJobs(tableId))
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(finishedThenCommittedStatusWithDefaults(job));
+                .containsExactly(compactionJobCreated(job, ignoredUpdateTime(),
+                        startedCompactionRun(DEFAULT_TASK_ID, defaultStartTime()),
+                        finishedCompactionRun(DEFAULT_TASK_ID, defaultSummary(), defaultCommitTime())));
     }
 }
