+++ java/compaction/compaction-tracker/src/test/java/sleeper/compaction/tracker/job/StoreCompactionJobRunIdIT.java
@@ -13,28 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.trackerv2.job;
+package sleeper.compaction.tracker.job;
 
 import org.junit.jupiter.api.Test;
 
 import sleeper.compaction.core.job.CompactionJob;
 import sleeper.compaction.core.job.CompactionJobStatusFromJobTestData;
-import sleeper.compaction.trackerv2.testutils.DynamoDBCompactionJobTrackerTestBase;
+import sleeper.compaction.tracker.testutils.DynamoDBCompactionJobTrackerTestBase;
 import sleeper.core.partition.Partition;
 import sleeper.core.statestore.FileReferenceFactory;
-import sleeper.core.tracker.job.run.JobRunSummary;
-import sleeper.core.tracker.job.run.RecordsProcessed;
 
-import java.time.Instant;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.finishedCompactionRun;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.compactionCommittedStatus;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.compactionFinishedStatus;
+import static sleeper.core.tracker.compaction.job.CompactionJobStatusTestData.compactionStartedStatus;
+import static sleeper.core.tracker.job.run.JobRunTestData.jobRunOnTask;
 
-public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTestBase {
+public class StoreCompactionJobRunIdIT extends DynamoDBCompactionJobTrackerTestBase {
 
     @Test
-    public void shouldReportCompactionJobStarted() {
+    public void shouldReportStartedJob() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
@@ -44,7 +44,8 @@ public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTes
 
         // When
         storeJobCreated(job);
-        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
 
         // Then
         assertThat(getAllJobStatuses())
@@ -53,7 +54,7 @@ public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTes
     }
 
     @Test
-    public void shouldReportCompactionJobUncommitted() {
+    void shouldReportUncommittedJob() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
@@ -63,8 +64,10 @@ public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTes
 
         // When
         storeJobCreated(job);
-        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
 
         // Then
         assertThat(getAllJobStatuses())
@@ -73,7 +76,7 @@ public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTes
     }
 
     @Test
-    public void shouldReportCompactionJobFinished() {
+    public void shouldReportFinishedAndCommittedJob() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
@@ -83,9 +86,12 @@ public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTes
 
         // When
         storeJobCreated(job);
-        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobFinished(job.finishedEventBuilder(defaultSummary()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
+        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
 
         // Then
         assertThat(getAllJobStatuses())
@@ -94,60 +100,62 @@ public class StoreCompactionJobUpdatesIT extends DynamoDBCompactionJobTrackerTes
     }
 
     @Test
-    public void shouldReportCompactionJobFailed() {
+    public void shouldReportFailedJob() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
         CompactionJob job = jobFactory.createCompactionJob(
                 List.of(fileFactory.rootFile(100L)),
                 partition.getId());
-        List<String> failureReasons = List.of("Something went wrong");
 
         // When
         storeJobCreated(job);
-        tracker.jobStarted(job.startedEventBuilder(defaultStartTime()).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
-        tracker.jobFailed(job.failedEventBuilder(defaultFinishTime()).failureReasons(failureReasons).taskId(DEFAULT_TASK_ID).jobRunId("test-run").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime())
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
+        tracker.jobFailed(job.failedEventBuilder(defaultFinishTime()).failure(new RuntimeException("Failed"))
+                .taskId(DEFAULT_TASK_ID).jobRunId("test-job-run").build());
 
         // Then
         assertThat(getAllJobStatuses())
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
-                .containsExactly(failedStatusWithDefaults(job, failureReasons));
+                .containsExactly(failedStatusWithDefaults(job, List.of("Failed")));
     }
 
     @Test
-    public void shouldReportLatestUpdatesWhenJobIsRunMultipleTimes() {
+    public void shouldReportJobFinishedTwiceOnSameTaskThenCommittedTwice() {
         // Given
         Partition partition = singlePartition();
         FileReferenceFactory fileFactory = fileFactory(partition);
         CompactionJob job = jobFactory.createCompactionJob(
                 List.of(fileFactory.rootFile(100L)),
                 partition.getId());
-        Instant startTime1 = Instant.parse("2022-10-03T15:19:01.001Z");
-        Instant finishTime1 = Instant.parse("2022-10-03T15:19:31.001Z");
-        Instant commitTime1 = Instant.parse("2022-10-03T15:19:41.001Z");
-        Instant startTime2 = Instant.parse("2022-10-03T15:19:02.001Z");
-        Instant finishTime2 = Instant.parse("2022-10-03T15:19:32.001Z");
-        Instant commitTime2 = Instant.parse("2022-10-03T15:19:42.001Z");
-        RecordsProcessed processed = new RecordsProcessed(100L, 100L);
-        JobRunSummary summary1 = new JobRunSummary(processed, startTime1, finishTime1);
-        JobRunSummary summary2 = new JobRunSummary(processed, startTime2, finishTime2);
 
         // When
         storeJobCreated(job);
-        tracker.jobStarted(job.startedEventBuilder(startTime1).taskId(DEFAULT_TASK_ID).jobRunId("run-1").build());
-        tracker.jobStarted(job.startedEventBuilder(startTime2).taskId(DEFAULT_TASK_ID_2).jobRunId("run-2").build());
-        tracker.jobFinished(job.finishedEventBuilder(summary1).taskId(DEFAULT_TASK_ID).jobRunId("run-1").build());
-        tracker.jobFinished(job.finishedEventBuilder(summary2).taskId(DEFAULT_TASK_ID_2).jobRunId("run-2").build());
-        tracker.jobCommitted(job.committedEventBuilder(commitTime1).taskId(DEFAULT_TASK_ID).jobRunId("run-1").build());
-        tracker.jobCommitted(job.committedEventBuilder(commitTime2).taskId(DEFAULT_TASK_ID_2).jobRunId("run-2").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime())
+                .taskId("test-task").jobRunId("test-run-1").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary())
+                .taskId("test-task").jobRunId("test-run-1").build());
+        tracker.jobStarted(job.startedEventBuilder(defaultStartTime())
+                .taskId("test-task").jobRunId("test-run-2").build());
+        tracker.jobFinished(job.finishedEventBuilder(defaultSummary())
+                .taskId("test-task").jobRunId("test-run-2").build());
+        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime())
+                .taskId("test-task").jobRunId("test-run-1").build());
+        tracker.jobCommitted(job.committedEventBuilder(defaultCommitTime())
+                .taskId("test-task").jobRunId("test-run-2").build());
 
         // Then
         assertThat(getAllJobStatuses())
                 .usingRecursiveFieldByFieldElementComparator(IGNORE_UPDATE_TIMES)
                 .containsExactly(CompactionJobStatusFromJobTestData.compactionJobCreated(job, ignoredUpdateTime(),
-                        finishedCompactionRun(DEFAULT_TASK_ID_2, new JobRunSummary(processed, startTime2, finishTime2),
-                                commitTime2),
-                        finishedCompactionRun(DEFAULT_TASK_ID, new JobRunSummary(processed, startTime1, finishTime1),
-                                commitTime1)));
+                        jobRunOnTask("test-task",
+                                compactionStartedStatus(defaultStartTime()),
+                                compactionFinishedStatus(defaultSummary()),
+                                compactionCommittedStatus(defaultCommitTime())),
+                        jobRunOnTask("test-task",
+                                compactionStartedStatus(defaultStartTime()),
+                                compactionFinishedStatus(defaultSummary()),
+                                compactionCommittedStatus(defaultCommitTime()))));
     }
 }
