+++ java/compaction/compaction-tracker/src/test/java/sleeper/compaction/tracker/task/DynamoDBCompactionTaskTrackerCreatorIT.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.compaction.trackerv2.task;
+package sleeper.compaction.tracker.task;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
