+++ java/compaction/compaction-tracker/src/test/java/sleeper/compaction/tracker/task/StoreCompactionTaskIT.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package sleeper.compaction.trackerv2.task;
+package sleeper.compaction.tracker.task;
 
 import org.junit.jupiter.api.Test;
 
-import sleeper.compaction.trackerv2.testutils.DynamoDBCompactionTaskTrackerTestBase;
+import sleeper.compaction.tracker.testutils.DynamoDBCompactionTaskTrackerTestBase;
 import sleeper.core.tracker.compaction.task.CompactionTaskStatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
