+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/compaction/AwsCompactionDriver.java
@@ -34,10 +34,10 @@ import sleeper.compaction.core.job.commit.CompactionCommitMessageSerDe;
 import sleeper.compaction.core.job.creation.CreateCompactionJobs;
 import sleeper.compaction.core.job.dispatch.CompactionJobDispatchRequest;
 import sleeper.compaction.core.job.dispatch.CompactionJobDispatchRequestSerDe;
-import sleeper.compaction.job.creationv2.AwsCreateCompactionJobs;
-import sleeper.compaction.job.creationv2.CompactionBatchJobsWriterToS3;
-import sleeper.compaction.job.creationv2.CompactionBatchMessageSenderToSqs;
-import sleeper.compaction.trackerv2.job.CompactionJobTrackerFactory;
+import sleeper.compaction.job.creation.AwsCreateCompactionJobs;
+import sleeper.compaction.job.creation.CompactionBatchJobsWriterToS3;
+import sleeper.compaction.job.creation.CompactionBatchMessageSenderToSqs;
+import sleeper.compaction.tracker.job.CompactionJobTrackerFactory;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.statestore.StateStoreProvider;
 import sleeper.core.tracker.compaction.job.CompactionJobTracker;
